Redo sample error calculations
The previous sample error estimates were wrong because they were calculating the SEM of the solution mean, while what we care about here is the SEM of the solution mean's error. These are subtly different, essentially SEM(true_mean - appx_mean) instead of SEM(true_mean) (where true_mean was approximated anyways). But since this is based on the error instead of just the end value, it doesn't make sense to do it by default in the WorkPrecision plots since the sample error line would only correspond to a single method. Thus it was scrapped from there.

But this means that get_sample_error had to be reworked so that way a setup can be run to get approximate solutions to then get the SEM of the statistical error.

The "true mean" is calculated as follows. If an analytical solution is known, a ton of draws of the final analytical solution is taken to take the mean. This should drown out the statistical noise, and is cheap since we can directly draw Brownians from tspan[2]. When there is no analytical solution, we take the mean of the means.

To get the sample error, we need to run the entire sampling process multiple times, and then take the standard deviation of the mean error. The total number of higher level times is controlled by sample_error_runs. This makes it super expensive, but this should be correct.